backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pillow
python-dotenv
django-cleanup